move the functions to the init module so that the main can be defined in poetry script format: `json_fnl:main`
